mouse input divide 3 classes
