+++ Assets/ZombieHunter/GameScene.unity
@@ -10810,7 +10810,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 75cdcd615ae7dd9408c135cf9ee5fe35, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  developmentMode: 0
+  developmentMode: 1
 --- !u!4 &1460235168
 Transform:
   m_ObjectHideFlags: 0
