+++ Assets/ZombieHunter/InputSystem/Resources.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: bdc29392b71270446be579d2e3c9a58f
-NativeFormatImporter:
+guid: f74f2ed37341436198c0f96cf8636946
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
-  mainObjectFileID: 11400000
   userData: 
   assetBundleName: 
   assetBundleVariant:
