+++ Assets/ZombieHunter/InputSystem/Resources/PlayerInputConfig.asset.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: a3228fa53df84f1c9fe2dab89b00a7a7
+timeCreated: 1646586351
