--- Assets/ZombieHunter/PlayerInputSystem/Resources.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 3feceea6c71c1654ebc7b55b61810842
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
