+++ Assets/ZombieHunter/WeaponSystem/WeaponSystemContainer.cs
@@ -1,5 +1,4 @@
 ﻿using Leopotam.Ecs;
-using UnityEngine;
 using ZombieHunter.WeaponSystem.Systems;
 
 namespace ZombieHunter.WeaponSystem
@@ -8,10 +7,10 @@ namespace ZombieHunter.WeaponSystem
     {
         public void AddSystems(EcsSystems ecsSystems)
         {
-            /*ecsSystems
+            ecsSystems
                 .Add(new BulletFlyingSystem())
                 .Add(new WeaponSpawnSystem())
-                .Add(new PlayerWeaponShootSystem());*/
+                .Add(new PlayerWeaponShootSystem());
         }
 
         public void AddInjectors(EcsSystems ecsSystems)
@@ -21,9 +20,9 @@ namespace ZombieHunter.WeaponSystem
 
         public void AddOneFrameObjects(EcsSystems ecsSystems)
         {
-            /*ecsSystems
+            ecsSystems
                 .OneFrame<ShootRightHandEvent>()
-                .OneFrame<ShootLeftHandEvent>();*/
+                .OneFrame<ShootLeftHandEvent>();
         }
     }
 }
