+++ UserSettings/EditorUserSettings.asset
@@ -24,13 +24,13 @@ EditorUserSettings:
       value: 224247031146466102030e2b150b4a1e02120a651b2d272022272e2af1f43135ade333e4e8751b39340da1290a371526
       flags: 0
     RecentlyUsedScenePath-6:
-      value: 224247031146466102030e2b150b4a1e02120a650b292b351e2a183de7ae2136ebf32f
-      flags: 0
-    RecentlyUsedScenePath-7:
       value: 224247031146466102030e2b150b4a1e02120a650127332328001323f7f40721f1f333e4a81e3a31364dc233112d0416fc1a1c1fd20e07045f1bf30705e6
       flags: 0
-    RecentlyUsedScenePath-8:
+    RecentlyUsedScenePath-7:
       value: 224247031146466102030e2b150b4a1e02120a6501273035202c1327d1f9272ce7ea79cde23730731d07e2334a2b0f36e613
+      flags: 0
+    RecentlyUsedScenePath-8:
+      value: 224247031146466102030e2b150b4a1e02120a650b292b351e2a183de7ae2136ebf32f
       flags: 0
     RecentlyUsedScenePath-9:
       value: 224247031146466102030e2b150b4a1e02120a65012733232805123ce9d32d2bf6e23ba6c33f3233762fe029173b2d30fd012d0efb04441e1f07e917
